fix: 주용이 의견 반영
